engine: extract esm/tours/tour_editor.js
